CRC: move monitor_signers to its own mod
